MorseCodeGenerator
* moved files to separate folder
* correctly file name for signal enum
